refactor FE testing for in-person-location-step & in-person-post-office-location-step
