+++ app/javascript/packages/document-capture/components/in-person-location-step.spec.tsx
@@ -69,97 +69,4 @@ describe('InPersonLocationStep', () => {
 
     await findByText('{"selected_location":"Baltimore"}');
   });
-
-  it('allows search by address when enabled', async () => {
-    const { findByText, findByLabelText } = render(
-      <InPersonContext.Provider value={{ arcgisSearchEnabled: true }}>
-        <InPersonLocationPostOfficeSearchStep {...DEFAULT_PROPS} />
-      </InPersonContext.Provider>,
-    );
-
-    await userEvent.type(
-      await findByLabelText('in_person_proofing.body.location.po_search.address_search_label'),
-      '100 main',
-    );
-    await userEvent.click(
-      await findByText('in_person_proofing.body.location.po_search.search_button'),
-    );
-    await findByText('in_person_proofing.body.location.po_search.results_description');
-  });
-
-  it('validates input and shows inline error', async () => {
-    const { findByText } = render(
-      <InPersonContext.Provider value={{ arcgisSearchEnabled: true }}>
-        <InPersonLocationPostOfficeSearchStep {...DEFAULT_PROPS} />
-      </InPersonContext.Provider>,
-    );
-
-    await userEvent.click(
-      await findByText('in_person_proofing.body.location.po_search.search_button'),
-    );
-
-    await findByText('in_person_proofing.body.location.inline_error');
-  });
-
-  it('displays no post office results if a successful search is followed by an unsuccessful search', async () => {
-    const { findByText, findByLabelText, queryByRole } = render(
-      <InPersonContext.Provider value={{ arcgisSearchEnabled: true }}>
-        <InPersonLocationPostOfficeSearchStep {...DEFAULT_PROPS} />
-      </InPersonContext.Provider>,
-    );
-
-    await userEvent.type(
-      await findByLabelText('in_person_proofing.body.location.po_search.address_search_label'),
-      '594 Broadway New York',
-    );
-    await userEvent.click(
-      await findByText('in_person_proofing.body.location.po_search.search_button'),
-    );
-
-    await userEvent.type(
-      await findByLabelText('in_person_proofing.body.location.po_search.address_search_label'),
-      'asdfkf',
-    );
-    await userEvent.click(
-      await findByText('in_person_proofing.body.location.po_search.search_button'),
-    );
-
-    const results = queryByRole('status', {
-      name: 'in_person_proofing.body.location.po_search.results_description',
-    });
-    expect(results).not.to.exist();
-  });
-
-  // TODO: fix below false positive test
-  it('displays a 500 error if the request to the USPS API throws an error', async () => {
-    const registerField = sandbox.stub();
-    const handleAddressFound = sandbox.stub();
-    const handleLocationsFound = sandbox.stub();
-    const handleError = (_: Error | null) => {};
-    const ADDRESS_SEARCH_PROPS = {
-      registerField,
-      onFoundAddress: handleAddressFound,
-      onFoundLocations: handleLocationsFound,
-      onError: handleError,
-    };
-    const { findByText, findByLabelText } = render(
-      <InPersonContext.Provider value={{ arcgisSearchEnabled: true }}>
-        <InPersonLocationPostOfficeSearchStep {...DEFAULT_PROPS}>
-          <AddressSearch {...ADDRESS_SEARCH_PROPS} />
-        </InPersonLocationPostOfficeSearchStep>
-      </InPersonContext.Provider>,
-    );
-
-    await userEvent.type(
-      await findByLabelText('in_person_proofing.body.location.po_search.address_search_label'),
-      '222 Merchandise Mart Plaza',
-    );
-
-    await userEvent.click(
-      await findByText('in_person_proofing.body.location.po_search.search_button'),
-    );
-
-    const error = findByText('idv.failure.exceptions.internal_error');
-    expect(error).to.exist();
-  });
 });
